moved Upload logic to ArchiveCommandBlocks
