+++ Arius/Arius.Core/Facade/Facade.cs
@@ -26,7 +26,6 @@ using System.Runtime.CompilerServices;
  * This is required to test the internals of the Arius.Core assembly
  */
 [assembly: InternalsVisibleTo("Arius.Core.Tests")]
-[assembly: InternalsVisibleTo("Arius.Core.Tests.Extensions")]
 [assembly: InternalsVisibleTo("Arius.Core.BehaviorTests")]
 
 namespace Arius.Core.Facade;
